requirements.txt
Flask==1.1.2
gevent==20.6.0
h5py==2.10.0
json5==0.9.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.18.1
pandas==1.0.3
Pillow==7.1.2
requests==2.23.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
gunicorn==20.0.4
opencv-python